Change 200 status code to 202 status code for non return resource function
