+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/json/nestedRecord.json
@@ -21,15 +21,8 @@
           }
         },
         "responses" : {
-          "200" : {
-            "description" : "Ok",
-            "content" : {
-              "text/plain" : {
-                "schema" : {
-                  "type" : "string"
-                }
-              }
-            }
+          "202" : {
+            "description" : "Accepted"
           }
         }
       }
