+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/path_scenario01.yaml
@@ -9,10 +9,6 @@ paths:
     post:
       operationId: operation1_post_/
       responses:
-        "200":
-          description: Ok
-          content:
-            text/plain:
-              schema:
-                type: string
+        "202":
+          description: Accepted
 components: {}
