+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/path_scenario02.yaml
@@ -9,19 +9,11 @@ paths:
     get:
       operationId: operation2_get_/
       responses:
-        "200":
-          description: Ok
-          content:
-            text/plain:
-              schema:
-                type: string
+        "202":
+          description: Accepted
     post:
       operationId: operation2_post_/
       responses:
-        "200":
-          description: Ok
-          content:
-            text/plain:
-              schema:
-                type: string
+        "202":
+          description: Accepted
 components: {}
