+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/path_scenario04.yaml
@@ -16,21 +16,13 @@ paths:
             type: integer
             format: int32
       responses:
-        "200":
-          description: Ok
-          content:
-            text/plain:
-              schema:
-                type: string
+        "202":
+          description: Accepted
     post:
       operationId: operation2_post_/
       responses:
-        "200":
-          description: Ok
-          content:
-            text/plain:
-              schema:
-                type: string
+        "202":
+          description: Accepted
   /v1/{id}/v2/{name}:
     get:
       operationId: "operation1_get_/v1/{id}/v2/{name}"
@@ -47,10 +39,6 @@ paths:
           schema:
             type: string
       responses:
-        "200":
-          description: Ok
-          content:
-            text/plain:
-              schema:
-                type: string
+        "202":
+          description: Accepted
 components: {}
