+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/request_body/json_payload.yaml
@@ -14,11 +14,7 @@ paths:
             schema:
               type: object
       responses:
-        "200":
-          description: Ok
-          content:
-            text/plain:
-              schema:
-                type: string
+        "202":
+          description: Accepted
 components: {}
