+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/request_body/mime_with_record_payload.yaml
@@ -14,12 +14,8 @@ paths:
             schema:
               $ref: '#/components/schemas/Pet'
       responses:
-        "200":
-          description: Ok
-          content:
-            text/plain:
-              schema:
-                type: string
+        "202":
+          description: Accepted
 components:
   schemas:
     Pet:
