+++ openapi-cli/src/test/resources/ballerina-to-openapi/expected_gen/request_body/multiple_payload.yaml
@@ -17,10 +17,6 @@ paths:
             schema:
               type: object
       responses:
-        "200":
-          description: Ok
-          content:
-            text/plain:
-              schema:
-                type: string
+        "202":
+          description: Accepted
 components: {}
